[REFACTORING] Split IMAPServerSSLConfigurationTest
